refactor: OrderConfirmPage 로직 분리
